+++ packages/headless/src/controllers/core/result-list/headless-core-result-list.ts
@@ -1,8 +1,9 @@
 import {ArrayValue, Schema, StringValue} from '@coveo/bueno';
 import {AsyncThunkAction} from '@reduxjs/toolkit';
 import {Result} from '../../../api/search/search/result';
+import {configuration} from '../../../app/common-reducers';
 import {CoreEngine} from '../../../app/engine';
-import {configuration, fields, search} from '../../../app/reducers';
+import {fields, search} from '../../../app/reducers';
 import {registerFieldsToInclude} from '../../../features/fields/fields-actions';
 import {
   ConfigurationSection,