+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.22.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.22.7...@coveo/quantic@2.22.8) (2023-05-11)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.22.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.22.6...@coveo/quantic@2.22.7) (2023-05-11)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.22.6](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.22.5...@coveo/quantic@2.22.6) (2023-05-09)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.22.5](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.22.4...@coveo/quantic@2.22.5) (2023-05-08)
+
+**Note:** Version bump only for package @coveo/quantic
+
+## [2.22.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.22.3...@coveo/quantic@2.22.4) (2023-05-05)
+
+**Note:** Version bump only for package @coveo/quantic
+
 ## [2.22.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@2.22.2...@coveo/quantic@2.22.3) (2023-05-04)
 
 **Note:** Version bump only for package @coveo/quantic